pyproject.toml
[project]
name = "openvote"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "eth-ape[recommended-plugins]>=0.8.32",
]
